requirements.txt
# ============================================
# H-OOI: Automated Insight Engine
# Dependencies (Latest versions - pip handles resolution)
# ============================================

# Core Framework
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
python-dotenv
python-multipart

# Data Processing
pandas
openpyxl
numpy
matplotlib

# Database Connectors
sqlalchemy
pymongo

# LangChain Ecosystem (Latest)
langchain
langchain-core
langchain-openai
langchain-ollama

# LangGraph (Latest with prebuilt components)
langgraph
langgraph-prebuilt

# PDF Processing
pdfplumber

# Report Generation
python-pptx

# HTTP Client
httpx
aiohttp
